requirements.txt
tensorflow==2.5
tensorflow_addons==0.14.0
pandas==1.3.3
matplotlib==3.4.3
scikit_learn==1.0
pyyaml
tqdm